make calcBeacon an utility function
