+++ beacon/handlers.go
@@ -156,11 +156,9 @@ func (pd *ProtocolDriver) classifyProposalMessage(ctx context.Context, m Proposa
 	case pd.isValidProposalMessage(currentEpoch, atxTimestamp, nextEpochStart, receivedTime):
 		logger.Debug("received valid proposal message")
 		pd.addValidProposal(m.VRFSignature)
-
 	case pd.isPotentiallyValidProposalMessage(currentEpoch, atxTimestamp, nextEpochStart, receivedTime):
 		logger.Debug("received potentially valid proposal message")
 		pd.addPotentiallyValidProposal(m.VRFSignature)
-
 	default:
 		logger.Warning("received invalid proposal message")
 	}
@@ -485,8 +483,8 @@ func (pd *ProtocolDriver) storeFollowingVotes(message FollowingVotingMessage, mi
 		}
 	}
 
-	// TODO(kimmy): keep later rounds votes in a separate buffer so we don't count them prematurely
-	// tho i am not sure whether counting votes for later rounds early is a security concern.
+	// TODO: don't accept votes in future round
+	// https://github.com/spacemeshos/go-spacemesh/issues/2794
 	for vote := range thisRoundVotes.invalid {
 		if _, ok := pd.votesMargin[vote]; !ok {
 			pd.votesMargin[vote] = new(big.Int).Neg(voteWeight)
